# Conflicts:
#	EmploymentWorks.vcxproj.filters
#	Game/Event/GoalBox.cpp
#	Game/Event/GoalBoxPortalEffect.cpp
#	Game/Event/LineCpuEffect.cpp
#	Game/Event/PortalFlame.cpp
#	Game/Game.cpp
#	KazLibrary/Buffer/CreateGpuBuffer.cpp
#	KazLibrary/Buffer/CreateGpuBuffer.h
#	KazLibrary/DirectXCommon/Base.h
#	KazLibrary/Helper/KazImGuiHelper.h
#	KazLibrary/Render/CircleRender.cpp
#	KazLibrary/Render/CircleRender.h
#	KazLibrary/Scene/PortalScene.cpp
#	KazLibrary/Scene/PortalScene.h
#	KazLibrary/Scene/SceneManager.cpp

+++ Game/Interface/IEnemy.h
@@ -103,7 +103,7 @@ public:
 	/// <summary>
 	/// €–S‰‰o
 	/// </summary>
-	void DeadEffect(KazMath::Vec3<float> *POS, KazMath::Vec3<float> *ROTATION, float *ALPHA);
+	void DeadEffect(KazMath::Vec3<float> *POS, KazMath::Vec3<float> *ROTATION, int *ALPHA);
 
 	
 	/// <returns>“G‚Ìî•ñ</returns>
